refactor: common fed test fn
