+++ 2014_ECRTS_federated_original/src/federated.rs
@@ -112,59 +112,8 @@ pub fn federated(dag_set: &mut [Graph<NodeData, i32>], number_of_cores: usize) -
 #[cfg(test)]
 mod tests {
     use super::*;
-    use lib::tests_helper::create_node;
+    use lib::tests_helper::{create_high_utilization_dag, create_low_utilization_dag};
     use std::collections::BTreeMap;
-
-    fn create_high_utilization_dag() -> Graph<NodeData, i32> {
-        let mut dag = Graph::<NodeData, i32>::new();
-        let n0 = {
-            let mut params = BTreeMap::new();
-            params.insert("execution_time".to_owned(), 4);
-            params.insert("period".to_owned(), 10);
-            dag.add_node(NodeData { id: 3, params })
-        };
-        let n1 = dag.add_node(create_node(1, "execution_time", 4));
-        let n2 = dag.add_node(create_node(2, "execution_time", 3));
-        let n3 = dag.add_node(create_node(3, "execution_time", 3));
-        dag.add_edge(n0, n1, 1);
-        dag.add_edge(n0, n2, 1);
-        dag.add_edge(n0, n3, 1);
-
-        dag
-    }
-
-    fn create_low_utilization_dag() -> Graph<NodeData, i32> {
-        let mut dag = Graph::<NodeData, i32>::new();
-        let n0 = {
-            let mut params = BTreeMap::new();
-            params.insert("execution_time".to_owned(), 3);
-            params.insert("period".to_owned(), 30);
-            dag.add_node(NodeData { id: 2, params })
-        };
-        let n1 = dag.add_node(create_node(0, "execution_time", 3));
-        let n2 = dag.add_node(create_node(1, "execution_time", 4));
-
-        dag.add_edge(n0, n1, 1);
-        dag.add_edge(n0, n2, 1);
-        dag
-    }
-
-    fn create_period_exceeding_dag() -> Graph<NodeData, i32> {
-        let mut dag = Graph::<NodeData, i32>::new();
-        let mut params = BTreeMap::new();
-        params.insert("execution_time".to_owned(), 20);
-        params.insert("period".to_owned(), 10);
-        dag.add_node(NodeData { id: 0, params });
-        dag
-    }
-
-    fn create_no_has_period_dag() -> Graph<NodeData, i32> {
-        let mut dag = Graph::<NodeData, i32>::new();
-        let mut params = BTreeMap::new();
-        params.insert("execution_time".to_owned(), 3);
-        dag.add_node(NodeData { id: 0, params });
-        dag
-    }
 
     #[test]
     fn test_federated_enough_core() {
@@ -219,8 +168,13 @@ mod tests {
 
     #[test]
     fn test_federated_unsuited_tasks() {
+        let mut dag = Graph::<NodeData, i32>::new();
+        let mut params = BTreeMap::new();
+        params.insert("execution_time".to_owned(), 20);
+        params.insert("period".to_owned(), 10);
+        dag.add_node(NodeData { id: 0, params });
         assert_eq!(
-            federated(&mut [create_period_exceeding_dag()], 5),
+            federated(&mut [dag], 5),
             Unschedulable {
                 reason: (String::from(
                     "The critical path length is greater than end_to_end_deadline."
@@ -233,6 +187,10 @@ mod tests {
     #[test]
     #[should_panic]
     fn test_federated_no_has_period() {
-        federated(&mut [create_no_has_period_dag()], 1);
+        let mut dag = Graph::<NodeData, i32>::new();
+        let mut params = BTreeMap::new();
+        params.insert("execution_time".to_owned(), 3);
+        dag.add_node(NodeData { id: 0, params });
+        federated(&mut [dag], 1);
     }
 }
